split Preprocessing module, update default config, update additional keywords, and add unit tests
