+++ src/dackar/config/nlp_config_default.toml
@@ -14,5 +14,8 @@ time_keywords_file = "../../../data/time_keywords.csv"
 time_relation_file = "../../../data/time_relation_keywords.csv"
 alias_file = "../../../data/alias.csv"
 
+extra_vocab = "../../../data/extra_vocab.txt"
 
+[params]
+spacy_language_pipeline = "en_core_web_lg"
